[Offload] refactoring ...
