+++ aten/src/ATen/native/hammerblade/Offload.h
@@ -1,43 +1,17 @@
+#pragma once
+
 #include <ATen/ATen.h>
 #include <ATen/Context.h>
 #include <ATen/Dispatch.h>
 #include <ATen/hammerblade/HammerBladeContext.h>
 #include <ATen/native/TensorIterator.h>
 
+#include <ATen/native/hammerblade/OffloadDef.h>
+#include <ATen/native/hammerblade/OffloadUtils.h>
+#include <ATen/native/hammerblade/OffloadIter.h>
+
 namespace at {
 namespace native {
-
-/**
- * HammerBlade Tensor Struct
- *
- * This struct defines the tensor layout on HB device.
- * HB kernel offloading routines cast PyTorch's tensors
- * to this format before loading and launching a kernel.
- * The layout of this struct matches that of the C struct
- * defined in HB device runtime.
- */
-typedef struct {
-  uint32_t N;    // Number of elements in the tensor
-  uint32_t dims; // Number of dimensions
-  eva_t strides; // Pointer to stride vector; number of strides = dims
-  eva_t data;    // Pointer to raw data
-} hb_mc_tensor_t;
-
-/**
- * HammerBlade Vector
- */
-typedef struct {
-  uint32_t N;
-  eva_t data;
-} hb_mc_vector_t;
-
-void offload_op_nullary(TensorIterator& iter, Scalar alpha, const char* kernel);
-void offload_op_unary(TensorIterator& iter, Scalar alpha, const char* kernel);
-void offload_op_binary(TensorIterator& iter, Scalar alpha, const char* kernel);
-
-void offload_op_nullary(TensorIterator& iter, Scalar beta, Scalar alpha, const char* kernel);
-void offload_op_unary(TensorIterator& iter, Scalar beta, Scalar alpha, const char* kernel);
-void offload_op_binary(TensorIterator& iter, Scalar beta, Scalar alpha, const char* kernel);
 
 void offload_memcpy(eva_t dest, eva_t src, uint32_t n);
 
@@ -96,7 +70,7 @@ do {                                                                            
   std::vector<Scalar> scalars;                                                                          \
   offload_tensor_scalar_impl(args, scalars, kernel);                                                    \
 } while (0);
-       
+
 void offload_convolution_forward(Tensor& output, const Tensor& input,
     const Tensor& weight, IntArrayRef padding, IntArrayRef stride,
     IntArrayRef dilation, int64_t groups);
